Updated App to control 2 LEDs

+++ MainScreen.js
@@ -3,27 +3,40 @@ import { useState } from 'react';
 import TcpSocket from 'react-native-tcp-socket';
 
 export default function MainScreen() {
-    const [led, setLed] = useState(0);
+    const [led_1, setLed_1] = useState(0);
+    const [led_2, setLed_2] = useState(0);
 
-    const [client, setClient] = useState(TcpSocket.createConnection(
-        {host: '192.168.4.1', port: 80}, 
-        () => {
+    const [client, setClient] = useState(
+        TcpSocket.createConnection({ host: '192.168.4.1', port: 80 },       () => {
             client.write('Hello World\r\n');
-        }
-    ));
+        })
+    );
 
-    const ledPress = () => {
+    const ledPress = (n) => {
+        const led = n == 1 ? led_1 : led_2;
+        const setLed = n == 1 ? setLed_1 : setLed_2;
+
         if (client.readyState == 'open') {
-            client.write(led == 0 ? 'LED:ON\r\n' : led == 1 ? 'LED:AUTO\r\n' : 'LED:OFF\r\n');
+            client.write(
+                led == 0
+                    ? `LED_${n}:ON\r\n`
+                    : led == 1
+                    ? `LED_${n}:AUTO\r\n`
+                    : `LED_${n}:OFF\r\n`
+            );
             setLed((led + 1) % 3);
         } else {
-            client.destroy();
-            setClient(TcpSocket.createConnection(
-                {host: '192.168.4.1', port: 80}, 
-                () => {
-                    client.write('Hello World\r\n');
-                }
-            ));
+            if (!client.destroyed) {
+                client.destroy();
+            }
+            setClient(
+                TcpSocket.createConnection(
+                    { host: '192.168.4.1', port: 80 },
+                    () => {
+                        client.write('Hello World\r\n');
+                    }
+                )
+            );
         }
 
         // fetch("http://192.168.117.1:8081/", {
@@ -41,14 +54,34 @@ export default function MainScreen() {
         <View style={styles.container}>
             <TouchableOpacity
                 style={
-                    led == 0
+                    led_1 == 0
                         ? [styles.button, styles.ledOff]
-                        : led == 1 ? [styles.button, styles.ledOn] : [styles.button, styles.ledauto]
+                        : led_1 == 1
+                        ? [styles.button, styles.ledOn]
+                        : [styles.button, styles.ledauto]
                 }
-                onPress={() => ledPress()}
+                onPress={() => ledPress(1)}
             >
-                <Text style={styles.heading}>LED</Text>
-                <Text style={styles.subheading}>{led == 0 ? 'OFF' : led == 1 ? 'ON' : 'AUTO'}</Text>
+                <Text style={styles.heading}>LED_1</Text>
+                <Text style={styles.subheading}>
+                    {led_1 == 0 ? 'OFF' : led_1 == 1 ? 'ON' : 'AUTO'}
+                </Text>
+            </TouchableOpacity>
+
+            <TouchableOpacity
+                style={
+                    led_2 == 0
+                        ? [styles.button, styles.ledOff]
+                        : led_2 == 1
+                        ? [styles.button, styles.ledOn]
+                        : [styles.button, styles.ledauto]
+                }
+                onPress={() => ledPress(2)}
+            >
+                <Text style={styles.heading}>LED_2</Text>
+                <Text style={styles.subheading}>
+                    {led_2 == 0 ? 'OFF' : led_2 == 1 ? 'ON' : 'AUTO'}
+                </Text>
             </TouchableOpacity>
         </View>
     );
